transcription-service/requirements.txt
psycopg[binary]==3.1.10
python-dotenv==1.0.0
fastapi
uvicorn[standard]
cohere